Better handling of "charset" property
Calling set_encoding method always marks the buffer as modified,
which is annoying and somewhat confusing.

Ideally the encoding would be configured before opening the file, but
there is no event for that, so we call io.reload_file() instead.

Fix test to handle this change correctly and rename _editorconfig
file to .editorconfig to remove the extra confname argument
to load_editorconfig().